version 3.1.1

+++ scala-st-std/src/main/scala/org.scalablytyped/std/RsaKeyAlgorithm.scala
@@ -6,12 +6,12 @@ import scala.scalajs.js.annotation._
 
 trait RsaKeyAlgorithm extends KeyAlgorithm {
   var modulusLength: Double
-  var publicExponent: Uint8Array
+  var publicExponent: BigInteger
 }
 
 object RsaKeyAlgorithm {
   @scala.inline
-  def apply(modulusLength: Double, name: java.lang.String, publicExponent: Uint8Array): RsaKeyAlgorithm = {
+  def apply(modulusLength: Double, name: java.lang.String, publicExponent: BigInteger): RsaKeyAlgorithm = {
     val __obj = js.Dynamic.literal(modulusLength = modulusLength.asInstanceOf[js.Any], name = name.asInstanceOf[js.Any], publicExponent = publicExponent.asInstanceOf[js.Any])
     __obj.asInstanceOf[RsaKeyAlgorithm]
   }
@@ -29,7 +29,7 @@ object RsaKeyAlgorithm {
     @scala.inline
     def setModulusLength(value: Double): Self = this.set("modulusLength", value.asInstanceOf[js.Any])
     @scala.inline
-    def setPublicExponent(value: Uint8Array): Self = this.set("publicExponent", value.asInstanceOf[js.Any])
+    def setPublicExponent(value: BigInteger): Self = this.set("publicExponent", value.asInstanceOf[js.Any])
   }
   
 }
